Massive commit of everything I have been doing over the last couple months, should have never gotten to this point.

+++ projects/TropixProteomicsCore/src/main/edu/umn/msi/tropix/proteomics/conversion/impl/MsmMgfScanWriterImpl.java
@@ -13,7 +13,7 @@ public class MsmMgfScanWriterImpl extends BaseMgfScanWriterImpl {
 
   @Override
   protected void writeTitle(final Scan scan, final short precursorCharge) {
-    final Float intensity = scan.getPrecursorIntensity();
+    final Double intensity = scan.getPrecursorIntensity();
     final String intensityString = (intensity > 0.0) ? String.format("precIntensity: %f ", intensity) : "";
     final String title = String.format("Elution from: 26.21 to 26.87 period: %s experiment: 1 cycles: 1 %sFinneganScanNumber %d",
         scan.getParentFileName(), intensityString, scan.getNumber());
